Use apt-get only on ubuntu

+++ docker.py
@@ -405,8 +405,9 @@ def make_dockerfile(conf):
                     dockerfile += 'RUN yum -y update && yum -y install lapack-devel && yum clean all\n'
             dockerfile += run_pip(req)
 
+    if 'ubuntu' in conf['base']:
+        dockerfile += 'RUN apt-get remove -y python3-six\n'
     # Make a user and home directory to install chainer
-    dockerfile += 'RUN apt-get remove -y python3-six\n'
     dockerfile += 'RUN useradd -m -u %d user\n' % os.getuid()
     return dockerfile
